requirements.txt
requests~=2.32.3
selenium~=4.21.0
pillow~=10.3.0